% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cypress.R
\name{add_sendtime2js}
\alias{add_sendtime2js}
\title{Add the sendTime function to the .js file}
\usage{
add_sendtime2js(js_file, txt_file)
}
\arguments{
\item{js_file}{Path to the .js file to add code}

\item{txt_file}{Path to the file to record the execution times}
}
\description{
Add the sendTime function to the .js file
}
\keyword{internal}
