% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkout}
\alias{checkout}
\title{Checkout GitHub branch}
\usage{
checkout(branch, debug)
}
\arguments{
\item{branch}{Commit hash code or branch name}

\item{debug}{Logical. TRUE to display all the system messages on runtime}
}
\description{
checkout and go to a different branch
}
\keyword{internal}
