% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkout_files}
\alias{checkout_files}
\title{Checkout GitHub files}
\usage{
checkout_files(debug)
}
\arguments{
\item{debug}{Logical. TRUE to display all the system messages on runtime}
}
\description{
Checkout anything created by the app. It prevents errors when
changing branches
}
\keyword{internal}
