% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{combine_performances}
\alias{combine_performances}
\title{Combine list of performances in a single data.frame}
\usage{
combine_performances(performance_list)
}
\arguments{
\item{performance_list}{list of tests containing commits with dates, duration
time and test name}
}
\description{
Combine list of performances in a single data.frame
}
