% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{command_wrapper}
\alias{command_wrapper}
\title{Wrapper to call on Operating System commands}
\usage{
command_wrapper(cmd, system = FALSE, intern = FALSE, ...)
}
\arguments{
\item{cmd}{command}

\item{system}{a logical (not NA) which indicates whether to use
shell or system call (system is a more low-level call)}

\item{intern}{a logical (not NA) which indicates whether to capture
the output of the command as an R character vector.}

\item{...}{Other paramters passed to shell or system}
}
\value{
see system or shell
}
\description{
Wrapper to call on Operating System commands
}
\keyword{internal}
