% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{commit_exists}
\alias{commit_exists}
\title{Check if git commit hash exists}
\usage{
commit_exists(commit)
}
\arguments{
\item{commit}{commit hash code, branch name or tag}
}
\value{
true if exists, an error message if not
}
\description{
Can be anything git recognizes as a commit, such
as a commit hash, a branch, a tag, ...
}
\keyword{internal}
