% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_cypress_list}
\alias{create_cypress_list}
\title{Create the cypress configuration list}
\usage{
create_cypress_list(plugins_file, port)
}
\arguments{
\item{plugins_file}{The path to the Cypress plugins}
}
\description{
Create the cypress configuration list
}
