% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cypress.R
\name{create_cypress_plugins}
\alias{create_cypress_plugins}
\title{Create the JS code to track execution time}
\usage{
create_cypress_plugins()
}
\description{
Create the JS code to track execution time
}
\keyword{internal}
