% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cypress.R
\name{create_cypress_structure}
\alias{create_cypress_structure}
\title{Create a temporary directory to store everything needed by Cypress}
\usage{
create_cypress_structure(app_dir, port, debug)
}
\arguments{
\item{app_dir}{The path to the application root}

\item{port}{Port to run the app}

\item{debug}{Logical. TRUE to display all the system messages on runtime}
}
\description{
Create a temporary directory to store everything needed by Cypress
}
\keyword{internal}
