% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cypress.R
\name{create_cypress_tests}
\alias{create_cypress_tests}
\title{Create the cypress files under project directory}
\usage{
create_cypress_tests(project_path, cypress_file)
}
\arguments{
\item{project_path}{The path to the project with all needed packages
installed}

\item{cypress_file}{The path to the .js file conteining cypress tests to
be recorded}
}
\description{
Create the cypress files under project directory
}
