% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cypress.R
\name{create_node_list}
\alias{create_node_list}
\title{Create the list of needed libraries}
\usage{
create_node_list(tests_path, port)
}
\arguments{
\item{tests_path}{The path to project}

\item{port}{Port to run the app}
}
\description{
Create the list of needed libraries
}
\keyword{internal}
