% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_progress_bar}
\alias{create_progress_bar}
\title{Create a progress bar to follow the execution}
\usage{
create_progress_bar(total = 100)
}
\arguments{
\item{total}{Total number of replications}
}
\description{
Create a progress bar to follow the execution
}
\keyword{internal}
