% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_shinytest2_structure}
\alias{create_shinytest2_structure}
\title{Create a temporary directory to store everything needed by shinytest2}
\usage{
create_shinytest2_structure(shinytest2_dir)
}
\arguments{
\item{shinytest2_dir}{The path to the shinytest2 tests}
}
\description{
Create a temporary directory to store everything needed by shinytest2
}
