% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_commit_date}
\alias{get_commit_date}
\title{Get the commit date in POSIXct format}
\usage{
get_commit_date(branch)
}
\arguments{
\item{branch}{Commit hash code or branch name}
}
\description{
Get the commit date in POSIXct format
}
\keyword{internal}
