% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shinytest2.R
\name{move_shinytest2_tests}
\alias{move_shinytest2_tests}
\title{Move tests to a temporary folder}
\usage{
move_shinytest2_tests(project_path, shinytest2_dir)
}
\arguments{
\item{project_path}{The path to the project}

\item{shinytest2_dir}{The directory with tests recorded by shinytest2}
}
\description{
Move tests to a temporary folder
}
\keyword{internal}
