% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{performance_test_cmd}
\alias{performance_test_cmd}
\title{Running the node script "performance_test" is system-dependent}
\usage{
performance_test_cmd(project_path)
}
\arguments{
\item{project_path}{path to project directory (one level above node)}
}
\description{
Running the node script "performance_test" is system-dependent
}
\keyword{internal}
