% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{prepare_file_paths}
\alias{prepare_file_paths}
\title{Prepare file paths for package and user sides report templates}
\usage{
prepare_file_paths(report_dir)
}
\arguments{
\item{report_dir}{name of the folder where the report should be saved}
}
\value{
two-element vector with paths to template reports
}
\description{
Prepare file paths for package and user sides report templates
}
\examples{
if (interactive()) {
  prepare_file_paths(report_dir = tempdir())
}
}
