% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.shiny_benchmark}
\alias{print.shiny_benchmark}
\title{Print for shiny_benchmark class}
\usage{
\method{print}{shiny_benchmark}(object)
}
\arguments{
\item{object}{shiny_benchmark object}
}
\description{
Print for shiny_benchmark class
}
