% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_cypress.R
\name{ptest_cypress}
\alias{ptest_cypress}
\title{Run the performance test based on multiple commits using Cypress}
\usage{
ptest_cypress(commit_list, cypress_file, app_dir, port, debug)
}
\arguments{
\item{commit_list}{A list of commit hash codes, branches' names or anything
else you can use with git checkout \link{...}}

\item{cypress_file}{The path to the .js file conteining cypress tests to
be recorded}

\item{app_dir}{The path to the application root}

\item{port}{Port to run the app}

\item{debug}{Logical. TRUE to display all the system messages on runtime}
}
\description{
Run the performance test based on multiple commits using Cypress
}
