% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_shinytest2.R
\name{ptest_shinytest2}
\alias{ptest_shinytest2}
\title{Run the performance test based on a multiple commits using shinytest2}
\usage{
ptest_shinytest2(commit_list, shinytest2_dir, app_dir, debug)
}
\arguments{
\item{commit_list}{A list of commit hash codes, branches' names or anything
else you can use with git checkout \link{...}}

\item{shinytest2_dir}{The directory with tests recorded by shinytest2
It can also be a vector of the same size of commit_list}

\item{app_dir}{The path to the application root}

\item{debug}{Logical. TRUE to display all the system messages on runtime}
}
\description{
Run the performance test based on a multiple commits using shinytest2
}
