% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore_env}
\alias{restore_env}
\title{Check and restore renv}
\usage{
restore_env(branch, renv_prompt)
}
\arguments{
\item{branch}{Commit hash code or branch name. Useful to create an
informative error message}

\item{renv_prompt}{Prompt the user before taking any action?}
}
\description{
Check whether renv is in use in the current branch. Raise error
if renv is not in use or apply renv:restore() in the case the package is
present
}
\keyword{internal}
