% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_tests.R
\name{run_performance_test}
\alias{run_performance_test}
\title{Run the performance test based on a single commit}
\usage{
run_performance_test(commit, project_path, cypress_file, txt_file, debug)
}
\arguments{
\item{commit}{A commit hash code or a branch's name}

\item{project_path}{The path to the project with all needed packages installed}

\item{cypress_file}{The path to the .js file conteining cypress tests to be recorded}

\item{txt_file}{The path to the file where it is aimed to save the times}

\item{debug}{Logical. TRUE to display all the system messages on runtime}
}
\description{
Run the performance test based on a single commit
}
