% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_tests.R
\name{run_shinytest2_performance_test}
\alias{run_shinytest2_performance_test}
\title{Run the performance test based on a single commit using shinytest2}
\usage{
run_shinytest2_performance_test(commit, app_dir, project_path, use_renv, debug)
}
\arguments{
\item{commit}{A commit hash code or a branch's name}

\item{project_path}{The path to the project with all needed packages installed}

\item{use_renv}{In case it is set as TRUE, package will try to apply
renv::restore() in all branches. Otherwise, the current loaded list of
packages will be used in all branches.}

\item{debug}{Logical. TRUE to display all the system messages on runtime}

\item{cypress_file}{The path to the .js file conteining cypress tests to be recorded}

\item{txt_file}{The path to the file where it is aimed to save the times}
}
\description{
Run the performance test based on a single commit using shinytest2
}
