% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_tests.R
\name{run_shinytest2_ptest}
\alias{run_shinytest2_ptest}
\title{Run the performance test based on a single commit using shinytest2}
\usage{
run_shinytest2_ptest(
  commit,
  app_dir,
  project_path,
  use_renv,
  renv_prompt,
  debug
)
}
\arguments{
\item{commit}{A commit hash code or a branch's name}

\item{app_dir}{The path to the application root}

\item{project_path}{The path to the project with all needed
packages installed}

\item{use_renv}{In case it is set as TRUE, package will try to apply
renv::restore() in all branches. Otherwise, the current loaded list of
packages will be used in all branches.}

\item{renv_prompt}{Prompt the user before taking any action?}

\item{debug}{Logical. TRUE to display all the system messages on runtime}
}
\description{
Run the performance test based on a single commit using shinytest2
}
