% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_benchmark-class.R
\docType{class}
\name{shiny_benchmark-class}
\alias{shiny_benchmark-class}
\alias{shiny_benchmark_class}
\title{An object of 'shiny_benchmark' class}
\description{
An object of 'shiny_benchmark' class
}
\section{Slots}{

\describe{
\item{\code{call}}{Function call}

\item{\code{time}}{Time elapsed}

\item{\code{performance}}{List of measurements (one entry for each commit)}
}}

