% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_commit}
\alias{summarise_commit}
\title{Return statistics based on the set of tests replications}
\usage{
summarise_commit(object)
}
\arguments{
\item{object}{A shiny_benchmark object}
}
\description{
Return statistics based on the set of tests replications
}
\keyword{internal}
