% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.shiny_benchmark}
\alias{summary.shiny_benchmark}
\title{Summary for shiny_benchmark class}
\usage{
\method{summary}{shiny_benchmark}(object, ...)
}
\arguments{
\item{object}{shiny_benchmark object}

\item{...}{Other parameters}
}
\value{
Return a \code{data.frame} with performance tests' summary statistics
}
\description{
Summary for shiny_benchmark class
}
