% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_app.R
\name{test_app_custom}
\alias{test_app_custom}
\title{Addapted from shinytest2::test_app}
\usage{
test_app_custom(
  app_dir = missing_arg(),
  tests_dir = fs::path(app_dir, "tests", "testthat"),
  ...,
  check_setup = TRUE
)
}
\description{
Addapted from shinytest2::test_app
}
