% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{combine_performances}
\alias{combine_performances}
\title{Combine list of performances in a single data.frame}
\usage{
combine_performances(performance_list)
}
\arguments{
\item{performance_list}{list of tests containing commits with dates, duration
time and test name}
}
\value{
data.frame combining all the elements from performance list. Also, it
adds the branch name for each one of the tables.
}
\description{
Combine list of performances in a single data.frame
}
\examples{
performance1 <- data.frame(t1 = 1, t2 = 2)
performance2 <- data.frame(t1 = 3, t2 = 4)
performance_list <- list(
  branch1 = performance1,
  branch2 = performance2
)

combine_performances(performance_list)

}
