% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cypress.R
\name{create_cypress_tests}
\alias{create_cypress_tests}
\title{Create the cypress files under project directory}
\usage{
create_cypress_tests(project_path, cypress_dir, tests_pattern)
}
\arguments{
\item{project_path}{The path to the project with all needed packages
installed}

\item{cypress_dir}{The directory with tests recorded by Cypress}

\item{tests_pattern}{Cypress files pattern. E.g. 'performance'. If it is NULL,
all the content will be used}
}
\description{
Create the cypress files under project directory
}
\keyword{internal}
