% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{create_report}
\alias{create_report}
\title{Create a performance report for the tests that were run}
\usage{
create_report(report_params, file = NULL)
}
\arguments{
\item{report_params}{list of tests containing commits with dates, duration
time and test name}

\item{file}{name of the file to which the report should be saved (.html)}
}
\description{
Create a performance report for the tests that were run
}
