% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{create_report}
\alias{create_report}
\title{Create a performance report for the tests that were run}
\usage{
create_report(report_params, file = NULL)
}
\arguments{
\item{report_params}{list of tests containing commits with dates, duration
time and test name}

\item{file}{name of the file to which the report should be saved (.html)}
}
\value{
None. This function is called for side effects

report_params <- list(
performance = data.frame(
date = Sys.time(),
rep_id = 1,
test_name = rep(c("t1", "t2"), each = 10),
duration_ms = rpois(n = 20, lambda = 10),
branch = paste0("b", 1:10)
)
)

create_report(
report_params = report_params,
file = tempfile(fileext = ".html")
)
}
\description{
Create a performance report for the tests that were run
}
