% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shinytest2.R
\name{create_shinytest2_structure}
\alias{create_shinytest2_structure}
\title{Create a temporary directory to store everything needed by shinytest2}
\usage{
create_shinytest2_structure(app_dir)
}
\arguments{
\item{app_dir}{The path to the application root}
}
\description{
Create a temporary directory to store everything needed by shinytest2
}
\keyword{internal}
