% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_example}
\alias{load_example}
\title{Load an application and instructions to run shiny.benchmark}
\usage{
load_example(path)
}
\arguments{
\item{path}{A character vector of full path names}
}
\description{
Load an application and instructions to run shiny.benchmark
}
