% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_tests.R
\name{performance_tests}
\alias{performance_tests}
\title{Execute performance tests for a list of commits}
\usage{
performance_tests(
  commit_list,
  cypress_file,
  app_dir = getwd(),
  port = 3333,
  debug = FALSE
)
}
\arguments{
\item{commit_list}{A list of commit hash codes, branches' names or anything else you can use with git checkout \link{...}}

\item{cypress_file}{The path to the .js file containing cypress tests to be recorded}

\item{app_dir}{The path to the application root}

\item{port}{Port to run the app}

\item{debug}{Logical. TRUE to display all the system messages on runtime}
}
\description{
Execute performance tests for a list of commits
}
