% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.shiny_benchmark}
\alias{plot.shiny_benchmark}
\title{Plot for shiny_benchmark class}
\usage{
\method{plot}{shiny_benchmark}(x, ...)
}
\arguments{
\item{x}{shiny_benchmark object}

\item{...}{Other parameters}
}
\value{
Return a \code{ggplot} object that compares different git refs
}
\description{
Plot for shiny_benchmark class
}
