% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_cypress.R
\name{run_cypress_ptest}
\alias{run_cypress_ptest}
\title{Run the performance test based on a single commit using Cypress}
\usage{
run_cypress_ptest(commit, project_path, cypress_file, debug)
}
\arguments{
\item{commit}{A commit hash code or a branch's name}

\item{project_path}{The path to the project with all needed packages
installed}

\item{cypress_file}{The path to the .js file conteining cypress tests to
be recorded}

\item{debug}{Logical. TRUE to display all the system messages on runtime}
}
\description{
Run the performance test based on a single commit using Cypress
}
