% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_tests.R
\name{run_shinytest2_ptest}
\alias{run_shinytest2_ptest}
\title{Run the performance test based on a single commit using shinytest2}
\usage{
run_shinytest2_ptest(commit, app_dir, project_path, debug)
}
\arguments{
\item{commit}{A commit hash code or a branch's name}

\item{app_dir}{The path to the application root}

\item{project_path}{The path to the project with all needed
packages installed}

\item{debug}{Logical. TRUE to display all the system messages on runtime}
}
\description{
Run the performance test based on a single commit using shinytest2
}
